meshlink_send()'s len argument should be size_t, to match send().

+++ src/meshlink.c
@@ -889,7 +889,7 @@ void meshlink_set_log_cb(meshlink_handle_t *mesh, meshlink_log_level_t level, me
 	mesh->log_level = level;
 }
 
-bool meshlink_send(meshlink_handle_t *mesh, meshlink_node_t *destination, const void *data, unsigned int len) {
+bool meshlink_send(meshlink_handle_t *mesh, meshlink_node_t *destination, const void *data, size_t len) {
 	if(!mesh || !destination)
 		return false;
 	if(!len)
